quantamend-vidora/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
jinja2==3.1.3
python-multipart==0.0.9
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
google-cloud-storage==2.15.0
python-dotenv==1.0.1
playwright==1.42.0
moviepy==1.0.3
pandas==2.2.1
